+++ volume/mounts/validate_unix_test.go
@@ -1,6 +1,6 @@
 // +build !windows
 
-package volume // import "github.com/docker/docker/volume"
+package mounts // import "github.com/docker/docker/volume/mounts"
 
 var (
 	testDestinationPath = "/foo"
