+++ volume/mounts/validate_windows_test.go
@@ -1,4 +1,4 @@
-package volume // import "github.com/docker/docker/volume"
+package mounts // import "github.com/docker/docker/volume/mounts"
 
 var (
 	testDestinationPath = `c:\foo`
